requirements.txt
Django>=3.1.7,<3.2.0
djangorestframework>=3.12.2,<3.13.0
djangorestframework-simplejwt>=5.1.0,<5.2.0
drf-yasg>=1.20.0,<1.21.0
